fix: rename model
