+++ src/app/book-page/book.service.ts
@@ -1,18 +1,18 @@
 import { Injectable } from '@angular/core';
 import bookStorage from './models/book-store';
-import {BookModel} from './models/book-model';
+import {Book} from './models/book-model';
 import {Observable, of} from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
 })
 export class BookService {
-  private books: BookModel[] = bookStorage;
+  private books: Book[] = bookStorage;
   private books$ = of(bookStorage);
 
   constructor() { }
 
-  getBooks(): Observable<BookModel[]> {
+  getBooks(): Observable<Book[]> {
     return this.books$;
   }
 }
