+++ src/app/core/constants.ts
@@ -1,0 +1,6 @@
+export enum BookCategories {
+  Action = 'Action',
+  Detective = 'Detective',
+  Fantasy = 'Fantasy',
+  Horror = 'Horror',
+}
